Added Frame to app
