+++ src/App.js
@@ -2,51 +2,16 @@ import React from 'react';
 
 import {BrowserRouter as Router, Route, Switch} from 'react-router-dom';
 
-import {makeStyles} from '@material-ui/core/styles';
-import AppBar from '@material-ui/core/AppBar';
-import Button from '@material-ui/core/Button';
-import Container from '@material-ui/core/Container'
-import Fab from '@material-ui/core/Fab';
-import Grid from '@material-ui/core/Grid';
-import IconButton from '@material-ui/core/IconButton';
-import Toolbar from '@material-ui/core/Toolbar';
-import Typography from '@material-ui/core/Typography';
-
-import MenuIcon from '@material-ui/icons/Menu';
-import KeyboardArrowUpIcon from '@material-ui/icons/KeyboardArrowUp'
+import Layout from './components/Layout'; 
 
 import Timekey from './outputs/timekey';
 import Market from './outputs/market';
 import Home from './outputs/home';
 
-const useStyles = makeStyles(theme => ({
-  root: {
-    flexGrow: 1,
-  },
-  menuButton: {
-    marginRight: theme.spacing(2),
-  },
-  title: {
-    flexGrow: 1,
-  },
-}));
-
 function App() {
-  const classes = useStyles();
+  
   return (
-    <>
-      <AppBar position="static">
-        <Toolbar>
-          <IconButton edge="start" className={classes.menuButton} color="inherit" aria-label="menu">
-            <MenuIcon />
-          </IconButton>
-          <Typography variant="h6" className={classes.title}>
-            A-NET Client Enterprise Application
-          </Typography>
-          <Button color="inherit">Login</Button>
-        </Toolbar>
-      </AppBar>
-      <Container fixed>
+    <Layout>
         <Router>
           <Switch>
             <Route exact path="/">
@@ -60,11 +25,7 @@ function App() {
             </Route>
           </Switch>
         </Router>
-      </Container>
-      <Fab color="secondary" size="small" style={{position: "absolute", bottom: '5px', right: '5px'}}>
-        <KeyboardArrowUpIcon />
-      </Fab>
-    </>
+      </Layout>
   );
 }
